+++ package.json
@@ -35,8 +35,7 @@
     "nodemon": "^1.18.7",
     "proxyquire": "^1.7.11",
     "sinon": "^2.4.1",
-    "supertest": "^3.3.0",
-    "supertest-as-promised": "^4.0.2"
+    "supertest": "^3.3.0"
   },
   "dependencies": {
     "cron": "^1.6.0",
